+++ arviz/data/io_emcee.py
@@ -115,15 +115,14 @@ class EmceeConverter:
 
     def posterior_to_xarray(self):
         """Convert the posterior to an xarray dataset."""
-        # A chain object is created outside of the dictionary to avoid a spike in memory usage
         # Use emcee3 syntax, else use emcee2
         if hasattr(self.sampler, "get_chain"):
-            chain = self.sampler.get_chain().swapaxes(0, 1)
+            samples_ary = self.sampler.get_chain().swapaxes(0, 1)
         else:
-            chain = self.sampler.chain
+            samples_ary = self.sampler.chain
 
         data = {
-            var_name: (chain[(..., idx)])
+            var_name: (samples_ary[(..., idx)])
             for idx, var_name in zip(self.slices, self.var_names)
         }
         return dict_to_dataset(